Add files via upload
These have the modifications I made to NtpClock to support the Nano 33 IOT which has built-in Wifi similar to the Maker1000 and Maker1010 borads, all of which are not supported by the current library because it insists that it work with an ESP32  or ESP8266.  The approach in my changes has been to just remove the restrictions in the case of the class definition, and to move the test to a place in the code where an attempt to connect to a wifi module occurs.  This didn't matter to me because I developed my own code to start up wifi, and just needed the clock functionality.  A more complete solution would either remove the test all together and deal with it somewhere else in the code as is done in the WiNNA examples.