Add files via upload
I modified OLED clock to work the GUI Slice library and Builder tool https://github.com/ImpulseAdventure/GUIslice/wiki and to support the Flash Storage library https://github.com/cmaglie/FlashStorage on the Arduino Nano 33 IOT.  It also uses the modified NtpClock library.  I haven't loaded the entire project here, just the key mods to the OLED code.   The change to the controller.h file centers around being able to manually set the clock if the Ntp service is not available (no wifi connection).   The change to the Presenter is to convert the output to strings rather than to the OLED and then export them via a callback.   The changes to the PersistentStore show a callback to read and write StoredInfo to the Flash storage.   Besides these changes, there is a GUI and a wifi manager that started out from the example in the WiFiNINA library WiFiUdpNtpClient, which shows a better alternative for testing for a valid wifi device than blocking as does the current NtpClock code.