ExtendedZoneProcessorTest: Fix syntax error on Seeed XIAO
The `const char* letter` was being initialized with a `{0}` because the
previous code defined `letter` as `char letter[2]`, and the test didn't get
updated. Weird thing is that both g++ and clang++ on Linux produced no errors,
nor did all the other Arduino compilers. It was only the ARM g++ for the SAMD21
that complained about this.