"zoneinfo/Brokers*.h: change ZoneRule.letter() to accept char[] buffer
Previously, ZoneRule.letter() returns either a `const char*` or `const
__FlashStringHelper`. But that forces the `BasicZoneProcessor` or
`ExtendedZoneProcessor` to decide whether to copy the `letter` from normal
memory or flash memory using `strncpy()` or `strncpy_P()`. This breaks the
template code so that it is no longer agnostic to whether the Zone database
comes from normal memory or flash memory. By making the caller supply the
`char[] buffer`, the persistence store knowledge can be fully encapsulated in
the broker classes.

A side-effect is that the `Transition` class no longer needs a `letter`
parameter (saving 2 or 4 bytes), because we can use the `abbrev` string buffer
to temporarily store the `letter` field, until `calcAbbreviation()` is called
to construct the actual abbreviation from the `format` and the `letter` (which
is temporarily stored in `abbrev`.